bubble sort
Define a function bubble_sort. It takes a list as an argument and sorts that list in place (rather than returning a new list).
optional challenges:
When you swap two list elements, instead of creating a temporary variable, take advantage of Python's multiple assignment feature.

Add an optional argument key that allows you to, for example, sort a list of strings either lexicographically or by length. It should work like the optional argument in the built-in sort function.

Suppose you had reason to believe that, in your application, the list passed to the function is often (but not always) already in reverse order. Provide a second version of the function that detects this situation and, when it occurs, sorts the list in linear time