implementation of Horspool's algorithm
motivation behind assignments:
This assignment challenges you to implement a sophisticated algorithm using the “transform” strategy. It is not immediately obvious that this algorithm solves the problem or that it does so efficiently. Getting this algorithm through your head with enough precision to produce a working implementation is a notable achievement.